fix: adjust prepare web server to init (#5277)
