+++ packages/server/core/src/base/middlewares/index.ts
@@ -1,3 +1,2 @@
-export * from './customServerHook';
-export * from './customServerMiddlewares';
 export * from './faviconFallback';
+export * from './customServer';
